refactor(dashboard): externalise le composant et modèle des statistiques dans des fichiers dédiés
- Déplace le template HTML de `DashboardStatsComponent` vers un fichier `dashboard-stats.html` séparé
- Introduit un fichier modèle pour définir l'interface `DashboardStats`
- Refactorise les chemins et imports pour homogénéiser la structure des fichiers du tableau de bord
- Simplifie les modules en renommant les classes et ajustant leurs responsabilités
