+++ src/app/features/dashboard/components/dashboard-stats/dashboard-stats.html
@@ -1,0 +1,26 @@
+<div class="grid grid-cols-2 md:grid-cols-3 lg:grid-cols-6 gap-4 pb-2">
+  <div class="bg-card border border-border rounded-xl p-4 text-center">
+    <div class="text-3xl font-bold text-text">{{ stats().total }}</div>
+    <div class="text-sm text-muted mt-1">Total</div>
+  </div>
+  <div class="bg-card border border-border rounded-xl p-4 text-center">
+    <div class="text-3xl font-bold text-info">{{ stats().applied }}</div>
+    <div class="text-sm text-muted mt-1">Envoyées</div>
+  </div>
+  <div class="bg-card border border-border rounded-xl p-4 text-center">
+    <div class="text-3xl font-bold text-warning">{{ stats().pending }}</div>
+    <div class="text-sm text-muted mt-1">En attente</div>
+  </div>
+  <div class="bg-card border border-border rounded-xl p-4 text-center">
+    <div class="text-3xl font-bold text-primary">{{ stats().interview }}</div>
+    <div class="text-sm text-muted mt-1">Entretiens</div>
+  </div>
+  <div class="bg-card border border-border rounded-xl p-4 text-center">
+    <div class="text-3xl font-bold text-success">{{ stats().accepted }}</div>
+    <div class="text-sm text-muted mt-1">Acceptées</div>
+  </div>
+  <div class="bg-card border border-border rounded-xl p-4 text-center">
+    <div class="text-3xl font-bold text-error">{{ stats().rejected }}</div>
+    <div class="text-sm text-muted mt-1">Refusées</div>
+  </div>
+</div>
