+++ src/app/features/dashboard/components/dashboard-stats/model/dashboard-stats.ts
@@ -1,0 +1,8 @@
+export interface DashboardStats {
+  total: number;
+  applied: number;
+  pending: number;
+  interview: number;
+  accepted: number;
+  rejected: number;
+}
